Started adding interrupts, already added basic NMI implementation and RTI command

+++ include/memory.h
@@ -5,13 +5,13 @@
 class memory {
 	public:
 		memory(void);
-		memory(uint16_t size);
+		memory(uint32_t size);
 		~memory(void);
 		uint8_t get_byte(uint16_t ind);
 		void set_byte(uint16_t ind, uint8_t val);
 		void read_file(std::string file);
 		void mem_dump(void);
 	private:
-		uint16_t sz;
+		uint32_t sz;
 		uint8_t* mem;
 };
